% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccle_crc_5}
\alias{ccle_crc_5}
\title{RNA-seq TPM data of 5 CRC cell line samples from CCLE.}
\format{
A DGEList of 19177 genes * 5 samples.
}
\source{
\code{\link[depmap:TPM]{depmap::depmap_TPM()}}
}
\usage{
data(ccle_crc_5)
}
\description{
A test DGEList object with RNA-seq RSEM quantified TPM data of 5 CRC cell
line samples from CCLE \code{\link[depmap:TPM]{depmap::depmap_TPM()}}.
}
\keyword{datasets}
