% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_functions.R
\name{DEGs_RP}
\alias{DEGs_RP}
\title{return DEGs UP and DOWN list based on Rank Product}
\usage{
DEGs_RP(
  tfit,
  lfc = 0,
  p = 0.05,
  assemble = "intersect",
  Rank = "adj.P.Val",
  nperm = 1e+05,
  keep.top = NULL,
  keep.group = NULL,
  ...
)
}
\arguments{
\item{tfit}{MArrayLM object generated by \code{\link[limma:ebayes]{limma::treat()}}}

\item{lfc}{num, cutoff of logFC for DE analysis}

\item{p}{num, cutoff of p value for DE analysis and permutation test if
method = "RP"}

\item{assemble}{'intersect' or 'union', whether to select intersected or
union genes of different comparisons, default 'intersect'}

\item{Rank}{chr, the variable for ranking DEGs, can be 'logFC', 'adj.P.Val'...
default 'adj.P.Val'}

\item{nperm}{num, permutation runs of simulating the distribution}

\item{keep.top}{NULL or num, whether to keep top n DEGs of specific comparison}

\item{keep.group}{NULL or pattern, specify the top DEGs of which comparison}

\item{...}{omitted
or group to be kept}
}
\description{
return DEGs UP and DOWN list based on Rank Product
}
