% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_non_tissue.R
\name{filter_non_tissue}
\alias{filter_non_tissue}
\alias{filter_non_tissue,matrix,vector,character-method}
\alias{filter_non_tissue,Matrix,vector,character-method}
\alias{filter_non_tissue,data.frame,vector,character-method}
\alias{filter_non_tissue,character,character,character-method}
\alias{filter_non_tissue,missing,character,character-method}
\alias{filter_non_tissue,DGEList,character,character-method}
\alias{filter_non_tissue,ExpressionSet,character,character-method}
\alias{filter_non_tissue,SummarizedExperiment,character,character-method}
\alias{filter_non_tissue,Seurat,character,character-method}
\title{Specify cell target_group signature against adherent cancer cell lines.}
\usage{
filter_non_tissue(
  data = "CCLE",
  group_col,
  target_group,
  ...,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ccle_tpm = NULL,
  ccle_meta = NULL
)

\S4method{filter_non_tissue}{matrix,vector,character}(
  data = "CCLE",
  group_col,
  target_group,
  ...,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ccle_tpm = NULL,
  ccle_meta = NULL
)

\S4method{filter_non_tissue}{Matrix,vector,character}(
  data = "CCLE",
  group_col,
  target_group,
  ...,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ccle_tpm = NULL,
  ccle_meta = NULL
)

\S4method{filter_non_tissue}{data.frame,vector,character}(
  data = "CCLE",
  group_col,
  target_group,
  ...,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ccle_tpm = NULL,
  ccle_meta = NULL
)

\S4method{filter_non_tissue}{character,character,character}(
  data = "CCLE",
  group_col,
  target_group,
  ...,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ccle_tpm = NULL,
  ccle_meta = NULL
)

\S4method{filter_non_tissue}{missing,character,character}(
  data = "CCLE",
  group_col,
  target_group,
  ...,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ccle_tpm = NULL,
  ccle_meta = NULL
)

\S4method{filter_non_tissue}{DGEList,character,character}(
  data = "CCLE",
  group_col,
  target_group,
  ...,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ccle_tpm = NULL,
  ccle_meta = NULL
)

\S4method{filter_non_tissue}{ExpressionSet,character,character}(
  data = "CCLE",
  group_col,
  target_group,
  ...,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ccle_tpm = NULL,
  ccle_meta = NULL
)

\S4method{filter_non_tissue}{SummarizedExperiment,character,character}(
  data = "CCLE",
  group_col,
  target_group,
  ...,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ccle_tpm = NULL,
  ccle_meta = NULL
)

\S4method{filter_non_tissue}{Seurat,character,character}(
  data = "CCLE",
  group_col,
  target_group,
  ...,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ccle_tpm = NULL,
  ccle_meta = NULL
)
}
\arguments{
\item{data}{'CCLE' or expression object}

\item{group_col}{vector or character, to specify the group of tumor/tissue
target_groups, or column name of \code{\link[depmap:metadata]{depmap::depmap_metadata()}},
e.g. 'primary_disease'}

\item{target_group}{pattern, specify the target_group of interest,
e.g. 'colorectal'}

\item{...}{params for \code{\link[=grep]{grep()}} to find matched cell lines in CCLE}

\item{log}{logical, if to do log transformation on expression}

\item{q}{num, quantile cutoff to screen markers which are not or lowly
expressed in relevant samples}

\item{markers}{vector, a vector of gene names, listed the gene symbols to be
filtered. Default 'NULL' means all genes}

\item{ignore.case}{logical, if to ignore the case of tissue pattern}

\item{slot}{character, specify which slot to use only for sce or seurat
object, optional, default 'counts'}

\item{ccle_tpm}{ccle_tpm data from \code{\link[depmap:TPM]{depmap::depmap_TPM()}}, only used when
data = 'CCLE', default NULL}

\item{ccle_meta}{ccle_meta data from \code{\link[depmap:metadata]{depmap::depmap_metadata()}}, only used
when data = 'CCLE', default NULL}
}
\value{
a vector of gene symbols
}
\description{
Specify signatures against specific tissues or cell lines,
and generate CCLE_tpm_new after retrieving specific tissues.
}
\examples{
data("NK_markers", "ccle_crc_5")
filter_non_tissue(ccle_crc_5, group_col = "cancer", target_group = "CRC",
                  markers = NK_markers$HGNC_Symbol)

}
