% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_subset_sig.R
\name{filter_subset_sig}
\alias{filter_subset_sig}
\alias{filter_subset_sig,list,vector-method}
\alias{filter_subset_sig,DGEList,vector-method}
\alias{filter_subset_sig,ANY,vector-method}
\title{Filter specific cell type signature genes against other subsets.}
\usage{
filter_subset_sig(
  data,
  markers,
  ID,
  type,
  counts = TRUE,
  dir = "UP",
  gene_id = "SYMBOL",
  method = "RP",
  comb = union,
  filter = c(10, 10),
  s_thres = 0.05,
  ...
)

\S4method{filter_subset_sig}{list,vector}(
  data,
  markers,
  ID,
  type,
  counts = TRUE,
  dir = "UP",
  gene_id = "SYMBOL",
  method = "RP",
  comb = union,
  filter = c(10, 10),
  s_thres = 0.05,
  ...
)

\S4method{filter_subset_sig}{DGEList,vector}(
  data,
  markers,
  ID,
  type,
  counts = TRUE,
  dir = "UP",
  gene_id = "SYMBOL",
  method = "RP",
  comb = union,
  filter = c(10, 10),
  s_thres = 0.05,
  ...
)

\S4method{filter_subset_sig}{ANY,vector}(
  data,
  markers,
  ID,
  type,
  counts = TRUE,
  dir = "UP",
  gene_id = "SYMBOL",
  method = "RP",
  comb = union,
  filter = c(10, 10),
  s_thres = 0.05,
  ...
)
}
\arguments{
\item{data}{A list of expression data objects}

\item{markers}{vector, a vector of gene names, listed the gene symbols to be
kept anyway after filtration. Default 'NULL' means no special
genes need to be kept.}

\item{ID}{vector of chr, specify the column name in coldata to compare}

\item{type}{vector of pattern, specify the cell type of interest}

\item{counts}{logical, if the expr in data is raw counts data}

\item{dir}{chr, could be 'UP' or 'DOWN' to use only up- or down-expressed
genes}

\item{gene_id}{chr, specify the gene ID type of rownames of expression data
when markers is not NULL, could be one of 'ENSEMBL', 'SYMBOL',
'ENTREZ'..., default 'SYMBOL'}

\item{method}{either 'RP' or 'GROUP', choose whether to use rank product or
group other subsets for multiple comparisons for DE analysis,
default 'RP'}

\item{comb}{'RRA' or Fun, keep all passing genes or only intersected genes,
could be \code{union} or \code{intersect} or \code{setdiff} or customed Fun,
or could be 'RRA' to use Robust Rank Aggregation method for
integrating multi-lists if DEGs, default 'union'}

\item{s_thres}{num, threshold of score if method = 'RRA'}

\item{...}{other params for \code{\link[=get_degs]{get_degs()}}}
}
\value{
a vector of gene symbols
}
\description{
Specify the signature of the subset matched 'type' against other subsets,
and generate proc_data after removing low expression genes and limma::voom
transformation.
}
\examples{
sigs <- filter_subset_sig(im_data_6, "celltype:ch1", "NK",
                          markers = NK_markers$HGNC_Symbol,
                          gene_id = "ENSEMBL")

}
