% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_table_generic.R
\name{get_DE_table}
\alias{get_DE_table}
\alias{get_DE_table,DGEList,character,character-method}
\alias{get_DE_table,matrix,vector,character-method}
\alias{get_DE_table,Matrix,vector,character-method}
\alias{get_DE_table,ExpressionSet,character,character-method}
\alias{get_DE_table,SummarizedExperiment,character,character-method}
\alias{get_DE_table,Seurat,character,character-method}
\title{Get DE analysis result table with statistics}
\usage{
get_DE_table(data, ID, type, slot = "counts", ...)

\S4method{get_DE_table}{DGEList,character,character}(data, ID, type, slot = "counts", ...)

\S4method{get_DE_table}{matrix,vector,character}(data, ID, type, slot = "counts", ...)

\S4method{get_DE_table}{Matrix,vector,character}(data, ID, type, slot = "counts", ...)

\S4method{get_DE_table}{ExpressionSet,character,character}(data, ID, type, slot = "counts", ...)

\S4method{get_DE_table}{SummarizedExperiment,character,character}(data, ID, type, slot = "counts", ...)

\S4method{get_DE_table}{Seurat,character,character}(data, ID, type, slot = "counts", ...)
}
\arguments{
\item{data}{expression object}

\item{ID}{chr, column name of coldata to specify the DE comparisons}

\item{type}{pattern, specify the group of interest, e.g. NK}

\item{slot}{chr, specify which slot to use only for sce or seurat object,
optional, default 'counts'}

\item{...}{params for function \code{\link[=DE_analysis]{DE_analysis()}}}
}
\value{
A list of DE result table of all comparisons.
}
\description{
This function uses edgeR and limma to get DE analysis results
lists for multiple comparisons. Filter out low expressed genes and obtain
DE statistics by using limma::voom and limma::treat, and also create an
object \code{proc_data} to store processed data.
}
\examples{
DE_tables <- get_DE_table(im_data_6, ID = "celltype:ch1", type = "NK")

}
