% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGs_generic.R
\name{get_DEGs}
\alias{get_DEGs}
\alias{get_DEGs,DGEList,character,character-method}
\alias{get_DEGs,matrix,vector,character-method}
\alias{get_DEGs,Matrix,vector,character-method}
\alias{get_DEGs,ExpressionSet,character,character-method}
\alias{get_DEGs,SummarizedExperiment,character,character-method}
\alias{get_DEGs,Seurat,character,character-method}
\title{Get differentially expressed genes by comparing specified groups}
\usage{
get_DEGs(
  data,
  ID,
  type,
  counts = TRUE,
  method = "RP",
  group = FALSE,
  filter = c(10, 10),
  plot = FALSE,
  lfc = 0,
  p = 0.05,
  nperm = 1e+05,
  rand = 123,
  assemble = "intersect",
  Rank = "adj.P.Val",
  markers = NULL,
  gene_id = "SYMBOL",
  keep.top = NULL,
  keep.group = NULL,
  slot = "counts"
)

\S4method{get_DEGs}{DGEList,character,character}(
  data,
  ID,
  type,
  counts = TRUE,
  method = "RP",
  group = FALSE,
  filter = c(10, 10),
  plot = FALSE,
  lfc = 0,
  p = 0.05,
  nperm = 1e+05,
  rand = 123,
  assemble = "intersect",
  Rank = "adj.P.Val",
  markers = NULL,
  gene_id = "SYMBOL",
  keep.top = NULL,
  keep.group = NULL
)

\S4method{get_DEGs}{matrix,vector,character}(
  data,
  ID,
  type,
  counts = TRUE,
  method = "RP",
  group = FALSE,
  filter = c(10, 10),
  plot = FALSE,
  lfc = 0,
  p = 0.05,
  nperm = 1e+05,
  rand = 123,
  assemble = "intersect",
  Rank = "adj.P.Val",
  markers = NULL,
  gene_id = "SYMBOL",
  keep.top = NULL,
  keep.group = NULL
)

\S4method{get_DEGs}{Matrix,vector,character}(
  data,
  ID,
  type,
  counts = TRUE,
  method = "RP",
  group = FALSE,
  filter = c(10, 10),
  plot = FALSE,
  lfc = 0,
  p = 0.05,
  nperm = 1e+05,
  rand = 123,
  assemble = "intersect",
  Rank = "adj.P.Val",
  markers = NULL,
  gene_id = "SYMBOL",
  keep.top = NULL,
  keep.group = NULL
)

\S4method{get_DEGs}{ExpressionSet,character,character}(
  data,
  ID,
  type,
  counts = TRUE,
  method = "RP",
  group = FALSE,
  filter = c(10, 10),
  plot = FALSE,
  lfc = 0,
  p = 0.05,
  nperm = 1e+05,
  rand = 123,
  assemble = "intersect",
  Rank = "adj.P.Val",
  markers = NULL,
  gene_id = "SYMBOL",
  keep.top = NULL,
  keep.group = NULL
)

\S4method{get_DEGs}{SummarizedExperiment,character,character}(
  data,
  ID,
  type,
  counts = TRUE,
  method = "RP",
  group = FALSE,
  filter = c(10, 10),
  plot = FALSE,
  lfc = 0,
  p = 0.05,
  nperm = 1e+05,
  rand = 123,
  assemble = "intersect",
  Rank = "adj.P.Val",
  markers = NULL,
  gene_id = "SYMBOL",
  keep.top = NULL,
  keep.group = NULL,
  slot = "counts"
)

\S4method{get_DEGs}{Seurat,character,character}(
  data,
  ID,
  type,
  counts = TRUE,
  method = "RP",
  group = FALSE,
  filter = c(10, 10),
  plot = FALSE,
  lfc = 0,
  p = 0.05,
  nperm = 1e+05,
  rand = 123,
  assemble = "intersect",
  Rank = "adj.P.Val",
  markers = NULL,
  gene_id = "SYMBOL",
  keep.top = NULL,
  keep.group = NULL,
  slot = "counts"
)
}
\arguments{
\item{data}{expression object}

\item{ID}{vector or chr, to specify the group of DE comparisons}

\item{type}{pattern, specify the group of interest, e.g. NK}

\item{counts}{logical, if the expr in data is raw counts data}

\item{method}{either 'RP' or 'Group', choose whether to use rank product or
group other subsets for multiple comparisons for DE analysis,
default 'RP'}

\item{group}{logical, TRUE to separate samples into only 2 groups:
`type`` and 'Others'; FALSE to set each level as a group}

\item{filter}{a vector of 2 numbers, filter condition to remove low expression
genes, the 1st for min.counts (if counts = TRUE) or CPM/TPM
(if counts = F), the 2nd for samples size 'large.n'}

\item{plot}{logical, if to make plots to show QC before and after filtration}

\item{lfc}{num, cutoff of logFC for DE analysis}

\item{p}{num, cutoff of p value for DE analysis and permutation test if
method = "RP"}

\item{nperm}{num, permutation runs of simulating the distribution}

\item{rand}{NULL or num, default 123, if NULL the permutation would be random}

\item{assemble}{'intersect' or 'union', whether to select intersected or
union genes of different comparisons, default 'intersect'}

\item{Rank}{chr, the variable for ranking DEGs, can be 'logFC', 'adj.P.Val'...
default 'adj.P.Val'}

\item{markers}{vector, a vector of gene names, listed the gene symbols to be
kept anyway after filtration. Default 'NULL' means no special
genes need to be kept.}

\item{gene_id}{chr, specify the gene ID type of rownames of expression data
when markers is not NULL, could be one of 'ENSEMBL', 'SYMBOL',
'ENTREZ'..., default 'SYMBOL'}

\item{keep.top}{NULL or num, whether to keep top n DEGs of specific comparison}

\item{keep.group}{NULL or pattern, specify the top DEGs of which comparison
or group to be kept}

\item{slot}{chr, specify which slot to use only for sce or seurat object,
optional, default 'counts'}
}
\value{
A list of 'UP' and 'DOWN' data frame of all differentially expressed
genes. Both are ordered by rank product or 'Rank' variable if
keep.top is NULL
}
\description{
This function uses edgeR and limma to get 'UP' and 'DOWN' DEG
lists, for multiple comparisons, DEGs can be obtained from intersection of
all comparsion DEGs or by using product of p value ranks for multiple
comparisons. Filter out low expressed genes and extract DE genes by using
limma::voom and limma::treat, and also create an object \code{proc_data} to store
processed data.
}
\examples{
DEGs <- get_DEGs(im_data_6, ID = "celltype:ch1",
                 type = "NK", gene_id = "ENSEMBL")

}
