% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LM.R
\name{get_LM}
\alias{get_LM}
\title{Extract specific subset markers from LM7 or/and LM22}
\usage{
get_LM(lm7.pattern = FALSE, lm22.pattern = FALSE, ...)
}
\arguments{
\item{lm7.pattern}{character string containing a regular expression
(or character string for fixed = TRUE), to be matched in
the given subsets in LM7}

\item{lm22.pattern}{character string containing a regular expression
(or character string for fixed = TRUE), to be matched in
the given subsets in LM22}

\item{...}{params for function \code{\link[=grep]{grep()}}}
}
\value{
A tibble with 2 or 3 columns: HGNC_Symbol and LM7/LM22
}
\description{
Extract specific subset markers from LM7 or/and LM22
}
\examples{
get_LM(lm7.pattern = "NK", lm22.pattern = "NK cells")
}
