% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lm_sig.R
\name{get_lm_sig}
\alias{get_lm_sig}
\title{Extract specific subset markers from LM7 or/and LM22}
\usage{
get_lm_sig(lm7.pattern, lm22.pattern, ...)
}
\arguments{
\item{lm7.pattern}{character string containing a regular expression,
to be matched in the given subsets in LM7}

\item{lm22.pattern}{character string containing a regular expression,
to be matched in the given subsets in LM22}

\item{...}{params for function \code{\link[=grep]{grep()}}}
}
\value{
A GeneSet or GeneSetCollection
}
\description{
Extract specific subset markers from LM7 or/and LM22
}
\examples{
get_lm_sig(lm7.pattern = "NK", lm22.pattern = "NK cells")
}
