% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Panglao.R
\name{get_Panglao}
\alias{get_Panglao}
\title{Extract immune subset markers from PanglaoDB website.}
\usage{
get_Panglao(species = "Hs", type)
}
\arguments{
\item{species}{chr, default 'Hs', could be 'Hs', 'Mm' or 'Mm Hs',
specify the species of interest}

\item{type}{character vector, cell type name(s) of interest,
available subsets could be listed by \code{\link[=list_Panglao_types]{list_Panglao_types()}}}
}
\value{
a tibble, first column is gene symbol, second column is markers origin
}
\description{
Extract specific immune subset markers for 'Hs' or 'Mm', the markers are
retrived from up-to-date PanglaoDB website.
}
\examples{
get_Panglao(species = "Hs", type = "NK cells")
}
