% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panglao_sig.R
\name{get_panglao_sig}
\alias{get_panglao_sig}
\title{Extract immune subset markers from PanglaoDB website.}
\usage{
get_panglao_sig(species = "Hs", type)
}
\arguments{
\item{species}{chr, default 'Hs', could be 'Hs', 'Mm' or 'Mm Hs',
specify the species of interest}

\item{type}{character vector, cell type name(s) of interest,
available subsets could be listed by \code{\link[=list_panglao_types]{list_panglao_types()}}}
}
\value{
a vector of markers
}
\description{
Extract specific immune subset markers for 'Hs' or 'Mm', the markers are
retrived from up-to-date PanglaoDB website.
}
\examples{
get_panglao_sig(species = "Hs", type = "NK cells")
}
