% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls2gsc.R
\name{gls2gsc}
\alias{gls2gsc}
\alias{gls2gsc,list-method}
\alias{gls2gsc,vector-method}
\title{Convert gene-set list into GeneSetCollection}
\usage{
gls2gsc(...)

\S4method{gls2gsc}{list}(...)

\S4method{gls2gsc}{vector}(...)
}
\arguments{
\item{...}{vector of genes or list of genes}
}
\value{
GeneSetCollection
}
\description{
Convert gene-set list into GeneSetCollection
}
\examples{
gls2gsc(GSEABase::geneIds(mastR::msigdb_gobp_nk[1:3]))
}
