% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsc_plot.R
\name{gsc_plot}
\alias{gsc_plot}
\title{Make upset plot for given gene sets}
\usage{
gsc_plot(...)
}
\arguments{
\item{...}{GeneSet or GeneSetCollection}
}
\value{
upset plot object
}
\description{
Plot upset diagram for overlapping genes among given gene-sets.
}
\examples{
gsc_plot(mastR::msigdb_gobp_nk[1:3])
}
