% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_panglao_organs.R
\name{list_panglao_organs}
\alias{list_panglao_organs}
\title{Show the summary info of available organs in PanglaoDB.}
\usage{
list_panglao_organs()
}
\value{
a vector of available organ types or cell types in PanglaoDB
}
\description{
Show the name of organs available in PanglaoDB. Help users know which organs
could be retrieved by PanglaoDB.
}
\examples{
list_panglao_organs()
}
