% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_panglao_types.R
\name{list_panglao_types}
\alias{list_panglao_types}
\title{Show the summary info of available cell types in PanglaoDB.}
\usage{
list_panglao_types(organ)
}
\arguments{
\item{organ}{character, specify the tissue or organ label to list cell types}
}
\value{
a vector of available cell types of the organ in PanglaoDB
}
\description{
Show the name and number of each cell type in PanglaoDB. Help users
know which subset(s) marker list(s) could be retrieved by PanglaoDB.
}
\examples{
list_panglao_types(organ = "Immune system")
}
