% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mastR-package.R
\docType{package}
\name{mastR_Package}
\alias{mastR-package}
\alias{mastR_Package}
\alias{mastR}
\title{Screen Immune Cells Signature for Specific Cancer or Tissue Type}
\value{
Automated screened signature
}
\description{
\code{mastR} This package enables automated screening of group specific signature
for specific tissues. The package is developed for generating refined lists
of signature genes from multiple group comparisons based on the results
from edgeR and limma differential expression (DE) analysis workflow. It
also takes into account the background expression of tissue-specificity,
which is often ignored by other markers generation tools. This package also
provides pseudo bulking function to deal with scRNA-seq data. Multiple
visualization functions are implemented in this package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://davislaboratory.github.io/mastR}
  \item Report bugs at \url{https://github.com/DavisLaboratory/mastR/issues}
}

}
\author{
Jinjin Chen \email{chen.j@wehi.edu.au}
}
\keyword{internal}
