% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_markers.R
\name{merge_markers}
\alias{merge_markers}
\title{Merge markers tbl into one.}
\usage{
merge_markers(markers_list, plot = FALSE)
}
\arguments{
\item{markers_list}{list of markers to be merged, can be list of vectors or
GeneSetCollection object}

\item{plot}{logical, if to make UpSetR plot for given list}
}
\value{
A data frame with genes symbols and columns indicate where the genes
come from
}
\description{
Merge markers collected from different DB into one tbl with 'TRUE' and '-' to
indicate which DB each gene is from.
}
\examples{
Msig <- get_msigdb_sig(
  pattern = "natural_killer_cell_mediated",
  ignore.case = TRUE
)
Panglao <- get_panglao_sig(type = "NK cells")
merge_markers(markers = list(NK_markers = NK_markers$HGNC_Symbol,
                             MSig = GSEABase::geneIds(Msig),
                             PanglaoDB = GSEABase::geneIds(Panglao)))
}
