% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msigdb_gobp_nk}
\alias{msigdb_gobp_nk}
\title{Sub-collection of MSigDB genesets.}
\format{
A GeneSetCollection of 55 genesets.
}
\source{
\code{\link[msigdb:getMsigdb]{msigdb::getMsigdb()}}
}
\usage{
msigdb_gobp_nk
}
\description{
A small GeneSetCollection object, contains genesets with geneset-name matched
to 'NATURAL_KILLER' from GO:BP MSigDB v7.4 database.
}
\keyword{datasets}
