% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mastR_Package.R
\docType{data}
\name{nk_markers}
\alias{nk_markers}
\title{NK cell markers combination.}
\format{
A data frame with 114 rows and  at least 4 variables:
\describe{
\item{HGNC_Symbol}{gene symbols}
\item{LM22}{if included in LM22}
\item{LM7}{if included in LM7}
\item{Huntington}{if included in orthologs}
}
}
\source{
\url{https://cancerimmunolres.aacrjournals.org/content/7/7/1162.long}
}
\usage{
data(nk_markers)
}
\value{
data frame
}
\description{
A dataset containing 114 NK cell markers from LM22,
LM7 and human orthologs in mice.
}
\keyword{datasets}
