% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_sig_by_step.R
\name{plot_mean_var}
\alias{plot_mean_var}
\title{plot Mean-variance trend after voom and after final linear fit}
\usage{
plot_mean_var(proc_data, span = 0.5)
}
\arguments{
\item{proc_data}{processed data returned by \code{\link[=process_data]{process_data()}}}

\item{span}{num, span for \code{\link[=lowess]{lowess()}}}
}
\value{
comparison plot of mean-variance of voom and final model
}
\description{
plot Mean-variance trend after voom and after final linear fit
}
\examples{
data("im_data_6")
proc_data <- process_data(
  im_data_6,
  group_col = "celltype:ch1",
  target_group = "NK"
)
plot_mean_var(proc_data)
}
