% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotPCAbiplot}
\alias{plotPCAbiplot}
\title{Single PCA plot function}
\usage{
plotPCAbiplot(
  prcomp,
  loading = FALSE,
  n_loadings = 10,
  dims = c(1, 2),
  group_by = NULL
)
}
\arguments{
\item{prcomp}{prcomp object generated by \code{\link[stats:prcomp]{stats::prcomp()}}}

\item{loading}{logical, if to plot and label loadings of PCA, default 'FALSE'}

\item{n_loadings}{num, top n loadings to plot; or a vector of gene IDs;
only work when \code{loading = TRUE}}

\item{dims}{a vector of 2 elements, specifying PCs to plot}

\item{group_by}{character, specify the column to be grouped and colored,
default NULL}
}
\value{
ggplot of PCA
}
\description{
Single PCA plot function
}
