% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mastr.R
\name{read_mastr}
\alias{read_mastr}
\title{read_mastr}
\usage{
read_mastr(
  csv_file = system.file("extdata/stromerzeuger_pv_berlin.csv", package = "mastr")
)
}
\arguments{
\item{csv_file}{defaut: system.file("extdata/stromerzeuger_pv_berlin.csv",
package = "mastr")}
}
\value{
imported mastr data
}
\description{
read_mastr
}
