% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_rankdensity_plot.R
\name{sig_rankdensity_plot}
\alias{sig_rankdensity_plot}
\alias{sig_rankdensity_plot,matrix,vector,vector-method}
\alias{sig_rankdensity_plot,Matrix,vector,vector-method}
\alias{sig_rankdensity_plot,data.frame,vector,vector-method}
\alias{sig_rankdensity_plot,DGEList,vector,character-method}
\alias{sig_rankdensity_plot,ExpressionSet,vector,character-method}
\alias{sig_rankdensity_plot,Seurat,vector,character-method}
\alias{sig_rankdensity_plot,SummarizedExperiment,vector,character-method}
\alias{sig_rankdensity_plot,list,vector,character-method}
\title{Plot rank density}
\usage{
sig_rankdensity_plot(
  data,
  sigs,
  ID,
  counts = TRUE,
  aggregate = FALSE,
  slot = "counts",
  gene_id = "SYMBOL"
)

\S4method{sig_rankdensity_plot}{matrix,vector,vector}(
  data,
  sigs,
  ID,
  counts = TRUE,
  aggregate = FALSE,
  gene_id = "SYMBOL"
)

\S4method{sig_rankdensity_plot}{Matrix,vector,vector}(
  data,
  sigs,
  ID,
  counts = TRUE,
  aggregate = FALSE,
  gene_id = "SYMBOL"
)

\S4method{sig_rankdensity_plot}{data.frame,vector,vector}(
  data,
  sigs,
  ID,
  counts = TRUE,
  aggregate = FALSE,
  gene_id = "SYMBOL"
)

\S4method{sig_rankdensity_plot}{DGEList,vector,character}(
  data,
  sigs,
  ID,
  counts = TRUE,
  aggregate = FALSE,
  gene_id = "SYMBOL"
)

\S4method{sig_rankdensity_plot}{ExpressionSet,vector,character}(
  data,
  sigs,
  ID,
  counts = TRUE,
  aggregate = FALSE,
  gene_id = "SYMBOL"
)

\S4method{sig_rankdensity_plot}{Seurat,vector,character}(
  data,
  sigs,
  ID,
  counts = TRUE,
  aggregate = FALSE,
  slot = "counts",
  gene_id = "SYMBOL"
)

\S4method{sig_rankdensity_plot}{SummarizedExperiment,vector,character}(
  data,
  sigs,
  ID,
  counts = TRUE,
  aggregate = FALSE,
  slot = "counts",
  gene_id = "SYMBOL"
)

\S4method{sig_rankdensity_plot}{list,vector,character}(
  data,
  sigs,
  ID,
  counts = TRUE,
  aggregate = FALSE,
  slot = "counts",
  gene_id = "SYMBOL"
)
}
\arguments{
\item{data}{expression data, can be matrix, DGEList, eSet, seurat, sce...}

\item{sigs}{a vector of signature (Symbols)}

\item{ID}{character or vector, specify the column name to compare in coldata}

\item{counts}{logical, indicate if data is raw counts data}

\item{aggregate}{logical, if to aggregate expression according to \code{ID},
default FALSE}

\item{slot}{character, indicate which slot used for seurat/sce object}

\item{gene_id}{character, indicate the ID type of rowname of expression data's ,
could be one of 'ENSEMBL', 'SYMBOL', ... default 'SYMBOL'}
}
\value{
ggplot or patchwork
}
\description{
Plot rank density
}
\examples{
sig_rankdensity_plot(
  data = im_data_6, sigs = NK_markers$HGNC_Symbol[1:10],
  ID = "celltype:ch1", gene_id = "ENSEMBL"
)

}
