% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TILsignaturesScreen_Package.R
\docType{package}
\name{TILsignaturesScreen_Package}
\alias{TILsignaturesScreen_Package}
\alias{TILsignaturesScreen}
\title{Screen Immune Cells Signature for Specific Cancer or Tissue Type}
\description{
\code{TILsignaturesScreen} This package enables screening immune cells signatures
for specific cancer or tissue type. Imported NK cell markers are based on
LM22 and LM7 and human orthologs in mice. MSigDB genesets of relevant terms
and Panglao DB markers can also be retrieved by this package. Users can
choose to use sorted immune cell bulk RNA-seq data and CCLE adherent cell
line RNA-seq data (CCLE) to screen signatures or use customized data.
This package also provides pseudo bulking function to deal with scRNA-seq data. Multiple
visualization functions are implemented in this package.
}
\author{
Jinjin Chen \email{chen.j@wehi.edu.au}
}
\keyword{internal}
