% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degs_generic.R
\name{de_analysis}
\alias{de_analysis}
\title{DE analysis pipeline}
\usage{
de_analysis(
  dge,
  ID,
  type,
  counts = TRUE,
  method = c("RP", "Group"),
  group = FALSE,
  filter = c(10, 10),
  plot = FALSE,
  lfc = 0,
  p = 0.05,
  markers = NULL,
  gene_id = "SYMBOL",
  ...
)
}
\arguments{
\item{dge}{DGEList object for DE analysis, including expr and samples info}

\item{ID}{character, column name of coldata to specify the DE comparisons}

\item{type}{pattern, specify the group of interest, e.g. NK}

\item{counts}{logical, if the expr in data is raw counts data}

\item{method}{either 'RP' or 'Group', choose whether to use rank product or
group other subsets for multiple comparisons for DE analysis,
default 'RP'}

\item{group}{logical, TRUE to separate samples into only 2 groups:
`type`` and 'Others'; FALSE to set each level as a group}

\item{filter}{a vector of 2 numbers, filter condition to remove low expression
genes, the 1st for min.counts (if counts = TRUE) or CPM/TPM
(if counts = F), the 2nd for samples size 'large.n'}

\item{plot}{logical, if to make plots to show QC before and after filtration}

\item{lfc}{num, cutoff of logFC for DE analysis}

\item{p}{num, cutoff of p value for DE analysis and permutation test if
method = "RP"}

\item{markers}{vector, a vector of gene names, listed the gene symbols to be
kept anyway after filtration. Default 'NULL' means no special
genes need to be kept.}

\item{gene_id}{character, specify the gene ID type of rownames of expression data
when markers is not NULL, could be one of 'ENSEMBL', 'SYMBOL',
'ENTREZ'..., default 'SYMBOL'}

\item{...}{omitted}
}
\value{
MArrayLM object generated by \code{\link[limma:ebayes]{limma::treat()}}
}
\description{
Standard DE analysis by using edgeR and limma::voom pipeline
}
