% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_non_tissue.R
\name{filter_non_tissue}
\alias{filter_non_tissue}
\alias{filter_non_tissue,matrix,vector,character-method}
\alias{filter_non_tissue,Matrix,vector,character-method}
\alias{filter_non_tissue,data.frame,vector,character-method}
\alias{filter_non_tissue,character,missing,character-method}
\alias{filter_non_tissue,missing,missing,character-method}
\alias{filter_non_tissue,DGEList,character,character-method}
\alias{filter_non_tissue,ExpressionSet,character,character-method}
\alias{filter_non_tissue,SummarizedExperiment,character,character-method}
\alias{filter_non_tissue,Seurat,character,character-method}
\title{Specify cell type signature against adherent cancer cell lines.}
\usage{
filter_non_tissue(
  data = "CCLE",
  ID,
  type,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ...
)

\S4method{filter_non_tissue}{matrix,vector,character}(
  data = "CCLE",
  ID,
  type,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ...
)

\S4method{filter_non_tissue}{Matrix,vector,character}(
  data = "CCLE",
  ID,
  type,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ...
)

\S4method{filter_non_tissue}{data.frame,vector,character}(
  data = "CCLE",
  ID,
  type,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ...
)

\S4method{filter_non_tissue}{character,missing,character}(
  data = "CCLE",
  ID,
  type,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ...
)

\S4method{filter_non_tissue}{missing,missing,character}(
  data = "CCLE",
  ID,
  type,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ...
)

\S4method{filter_non_tissue}{DGEList,character,character}(
  data = "CCLE",
  ID,
  type,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ...
)

\S4method{filter_non_tissue}{ExpressionSet,character,character}(
  data = "CCLE",
  ID,
  type,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ...
)

\S4method{filter_non_tissue}{SummarizedExperiment,character,character}(
  data = "CCLE",
  ID,
  type,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ...
)

\S4method{filter_non_tissue}{Seurat,character,character}(
  data = "CCLE",
  ID,
  type,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ...
)
}
\arguments{
\item{data}{'CCLE' or expression object}

\item{ID}{vector or chr, to specify the group of tumor/tissue types}

\item{type}{pattern, specify the type of interest, e.g. 'colorectal'}

\item{log}{logical, if to do log transformation on expression}

\item{q}{num, quantile cutoff to screen markers which are not or lowly
expressed in relevant samples}

\item{markers}{vector, a vector of gene names, listed the gene symbols to be
filtered. Default 'NULL' means all genes in data to be filtered}

\item{ignore.case}{logical, if to ignore the case of tissue pattern}

\item{slot}{chr, specify which slot to use only for sce or seurat object,
optional, default 'counts'}

\item{...}{params for \code{\link[=grep]{grep()}} to find matched cell lines in CCLE}
}
\value{
a vector of gene symbols
}
\description{
Specify signatures against specific tissues or cell lines,
and generate CCLE_tpm_new after retrieving specific tissues.
}
\examples{
filter_non_tissue(type = "colorectal", markers = NK_markers$HGNC_Symbol)

}
