% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_subset_sig.R
\name{filter_subset_sig}
\alias{filter_subset_sig}
\alias{filter_subset_sig,list-method}
\alias{filter_subset_sig,DGEList-method}
\alias{filter_subset_sig,ANY-method}
\title{Filter specific cell type signature genes against other subsets.}
\usage{
filter_subset_sig(
  data,
  ID,
  type,
  markers = NULL,
  counts = TRUE,
  dir = "UP",
  gene_id = "SYMBOL",
  method = "RP",
  comb = union,
  filter = c(10, 10),
  s_thres = 0.05,
  ...
)

\S4method{filter_subset_sig}{list}(
  data,
  ID,
  type,
  markers = NULL,
  counts = TRUE,
  dir = "UP",
  gene_id = "SYMBOL",
  method = "RP",
  comb = union,
  filter = c(10, 10),
  s_thres = 0.05,
  ...
)

\S4method{filter_subset_sig}{DGEList}(
  data,
  ID,
  type,
  markers = NULL,
  counts = TRUE,
  dir = "UP",
  gene_id = "SYMBOL",
  method = "RP",
  comb = union,
  filter = c(10, 10),
  s_thres = 0.05,
  ...
)

\S4method{filter_subset_sig}{ANY}(
  data,
  ID,
  type,
  markers = NULL,
  counts = TRUE,
  dir = "UP",
  gene_id = "SYMBOL",
  method = "RP",
  comb = union,
  filter = c(10, 10),
  s_thres = 0.05,
  ...
)
}
\arguments{
\item{data}{An expression data or a list of expression data objects}

\item{ID}{vector or character, specify the group factor or column name of
coldata for DE comparisons}

\item{type}{pattern, specify the group of interest, e.g. NK}

\item{markers}{vector, a vector of gene names, listed the gene symbols to be
kept anyway after filtration. Default 'NULL' means no special
genes need to be kept.}

\item{counts}{logical, if the expr in data is raw counts data}

\item{dir}{character, could be 'UP' or 'DOWN' to use only up- or
down-expressed genes}

\item{gene_id}{character, specify the gene ID type of rownames of expression data
when markers is not NULL, could be one of 'ENSEMBL', 'SYMBOL',
'ENTREZ'..., default 'SYMBOL'}

\item{method}{either 'RP' or 'Group', choose whether to use rank product or
group other subsets for multiple comparisons for DE analysis,
default 'RP'}

\item{comb}{'RRA' or Fun, keep all passing genes or only intersected genes,
could be \code{union} or \code{intersect} or \code{setdiff} or customed Fun,
or could be 'RRA' to use Robust Rank Aggregation method for
integrating multi-lists if DEGs, default 'union'}

\item{filter}{(list of) vector of 2 numbers, filter condition to remove low
expression genes, the 1st for min.counts (if counts = TRUE) or
CPM/TPM (if counts = F), the 2nd for samples size 'large.n'}

\item{s_thres}{num, threshold of score if comb = 'RRA'}

\item{...}{other params for \code{\link[=get_degs]{get_degs()}}}
}
\value{
a vector of gene symbols
}
\description{
Specify the signature of the subset matched 'type' against other subsets,
either "union", "intersect" or "RRA" can be specified when input is a list
of datasets to integrate the signatures into one.
}
\examples{
sigs <- filter_subset_sig(im_data_6, "celltype:ch1", "NK",
                          markers = NK_markers$HGNC_Symbol,
                          gene_id = "ENSEMBL")

}
