% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_table_generic.R
\name{get_de_table}
\alias{get_de_table}
\alias{get_de_table,DGEList,character,character-method}
\alias{get_de_table,matrix,vector,character-method}
\alias{get_de_table,Matrix,vector,character-method}
\alias{get_de_table,ExpressionSet,character,character-method}
\alias{get_de_table,SummarizedExperiment,character,character-method}
\alias{get_de_table,Seurat,character,character-method}
\title{Get DE analysis result table with statistics}
\usage{
get_de_table(data, ID, type, slot = "counts", ...)

\S4method{get_de_table}{DGEList,character,character}(data, ID, type, slot = "counts", ...)

\S4method{get_de_table}{matrix,vector,character}(data, ID, type, slot = "counts", ...)

\S4method{get_de_table}{Matrix,vector,character}(data, ID, type, slot = "counts", ...)

\S4method{get_de_table}{ExpressionSet,character,character}(data, ID, type, slot = "counts", ...)

\S4method{get_de_table}{SummarizedExperiment,character,character}(data, ID, type, slot = "counts", ...)

\S4method{get_de_table}{Seurat,character,character}(data, ID, type, slot = "counts", ...)
}
\arguments{
\item{data}{expression object}

\item{ID}{chr, column name of coldata to specify the DE comparisons}

\item{type}{pattern, specify the group of interest, e.g. NK}

\item{slot}{chr, specify which slot to use only for sce or seurat object,
optional, default 'counts'}

\item{...}{params for function \code{\link[=de_analysis]{de_analysis()}}}
}
\value{
A list of DE result table of all comparisons.
}
\description{
This function uses edgeR and limma to get DE analysis results
lists for multiple comparisons. Filter out low expressed genes and obtain
DE statistics by using limma::voom and limma::treat, and also create an
object \code{proc_data} to store processed data.
}
\examples{
DE_tables <- get_de_table(im_data_6, ID = "celltype:ch1", type = "NK")

}
