% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degs_generic.R
\name{get_degs}
\alias{get_degs}
\alias{get_degs,DGEList,character,character-method}
\alias{get_degs,matrix,vector,character-method}
\alias{get_degs,Matrix,vector,character-method}
\alias{get_degs,ExpressionSet,character,character-method}
\alias{get_degs,SummarizedExperiment,character,character-method}
\alias{get_degs,Seurat,character,character-method}
\title{Get differentially expressed genes by comparing specified groups}
\usage{
get_degs(
  data,
  ID,
  type,
  counts = TRUE,
  method = c("RP", "Group"),
  slot = "counts",
  ...
)

\S4method{get_degs}{DGEList,character,character}(
  data,
  ID,
  type,
  counts = TRUE,
  method = c("RP", "Group"),
  slot = "counts",
  ...
)

\S4method{get_degs}{matrix,vector,character}(
  data,
  ID,
  type,
  counts = TRUE,
  method = c("RP", "Group"),
  slot = "counts",
  ...
)

\S4method{get_degs}{Matrix,vector,character}(
  data,
  ID,
  type,
  counts = TRUE,
  method = c("RP", "Group"),
  slot = "counts",
  ...
)

\S4method{get_degs}{ExpressionSet,character,character}(
  data,
  ID,
  type,
  counts = TRUE,
  method = c("RP", "Group"),
  slot = "counts",
  ...
)

\S4method{get_degs}{SummarizedExperiment,character,character}(
  data,
  ID,
  type,
  counts = TRUE,
  method = c("RP", "Group"),
  slot = "counts",
  ...
)

\S4method{get_degs}{Seurat,character,character}(
  data,
  ID,
  type,
  counts = TRUE,
  method = c("RP", "Group"),
  slot = "counts",
  ...
)
}
\arguments{
\item{data}{expression object}

\item{ID}{vector or character, specify the group factor or column name of
coldata for DE comparisons}

\item{type}{pattern, specify the group of interest, e.g. NK}

\item{counts}{logical, if the expr in data is raw counts data}

\item{method}{either 'RP' or 'Group', choose whether to use rank product or
group other subsets for multiple comparisons for DE analysis,
default 'RP'}

\item{slot}{character, specify which slot to use only for sce or seurat
object, optional, default 'counts'}

\item{...}{params for \code{\link[=de_analysis]{de_analysis()}} and \code{\link[=DEGs_RP]{DEGs_RP()}} or \code{\link[=DEGs_Group]{DEGs_Group()}}}
}
\value{
A list of 'UP' and 'DOWN' data frame of all differentially expressed
genes. Both are ordered by rank product or 'Rank' variable if
keep.top is NULL
}
\description{
This function uses edgeR and limma to get 'UP' and 'DOWN' DEG
lists, for multiple comparisons, DEGs can be obtained from intersection of
all comparsion DEGs or by using product of p value ranks for multiple
comparisons. Filter out low expressed genes and extract DE genes by using
limma::voom and limma::treat, and also create an object \code{proc_data} to
store processed data.
}
\examples{
DEGs <- get_degs(im_data_6, ID = "celltype:ch1",
                 type = "NK", gene_id = "ENSEMBL")

}
