% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_msigdb_sig.R
\name{get_msigdb_sig}
\alias{get_msigdb_sig}
\title{Collect genes from MSigDB.}
\usage{
get_msigdb_sig(
  pattern,
  species = c("hs", "mm"),
  cat = NULL,
  subcat = NULL,
  id = c("SYM", "EZID"),
  version = msigdb::getMsigdbVersions(),
  plot = FALSE,
  ...
)
}
\arguments{
\item{pattern}{pattern pass to \code{\link[=grep]{grep()}}, to match the MsigDB gene-set name of
interest, e.g. 'NATURAL_KILLER_CELL_MEDIATED'}

\item{species}{character, species of interest, can be 'hs' or 'mm'}

\item{cat}{character, stating the category(s) to be retrieved.
The category(s) must be one from \code{\link[msigdb:listCollections]{msigdb::listCollections()}},
see details in \code{\link[msigdb:subsetCollection]{msigdb::subsetCollection()}}}

\item{subcat}{character, stating the sub-category(s) to be retrieved.
The sub-category(s) must be one from \code{\link[msigdb:listSubCollections]{msigdb::listSubCollections()}},
see details in \code{\link[msigdb:subsetCollection]{msigdb::subsetCollection()}}}

\item{id}{a character, representing the ID type to use ("SYM" for gene
symbols and "EZID" for Entrez IDs)}

\item{version}{a character, stating the version of MSigDB to be retrieved
(should be >= 7.2). See \code{\link[msigdb:getMsigdbVersions]{msigdb::getMsigdbVersions()}}.}

\item{plot}{logical, if to plot UpSetR diagram}

\item{...}{params for \code{\link[=grep]{grep()}}, used to match pattern to gs_name}
}
\value{
A GeneSet object containing all matched gene-sets in MSigDB
}
\description{
Collect genes of relevant MSigDB gene-sets, of which the gene-set names are
matched to the given pattern by using \code{\link[=grep]{grep()}} function.
}
\examples{
get_msigdb_sig(
  pattern = "natural_killer_cell_mediated",
  species = "hs", subcat = "GO:BP",
  version = '7.4',
  ignore.case = TRUE
)
}
