% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_msigdb_sig.R
\name{get_msigdb_sig}
\alias{get_msigdb_sig}
\title{Collect genes from MSigDB.}
\usage{
get_msigdb_sig(
  species = "Homo sapiens",
  cat = "C5",
  subcat = "GO:BP",
  pattern,
  plot = FALSE,
  ...
)
}
\arguments{
\item{species}{chr, species of interest, available species can be listed by
\code{\link[msigdbr:msigdbr_species]{msigdbr::msigdbr_species()}}}

\item{cat}{chr, category of interest, available collections can be listed by
\code{\link[msigdbr:msigdbr_collections]{msigdbr::msigdbr_collections()}}}

\item{subcat}{chr, subcategory of interest, available collections can be
listed by \code{\link[msigdbr:msigdbr_collections]{msigdbr::msigdbr_collections()}}}

\item{pattern}{pattern, to be matched the MsigDB gs_name of interest,
e.g. 'natural_killer_cell_mediated'}

\item{plot}{logical, if to plot UpSetR diagram}

\item{...}{params for \code{\link[=grep]{grep()}}, used to match pattern to gs_name}
}
\value{
A GeneSet object
}
\description{
Collect genes of relevant MSigDB genesets.
}
\examples{
get_msigdb_sig(
  species = "Homo sapiens", cat = "C5", subcat = "GO:BP",
  pattern = "natural_killer_cell_mediated", ignore.case = TRUE
)
}
