% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mastR-package.R
\docType{data}
\name{im_data_6}
\alias{im_data_6}
\title{RNA-seq TMM normalized counts data of 6 sorted immune subsets.}
\format{
An ExpressionSet objects of 6*4 samples.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE60424}
}
\usage{
data(im_data_6)
}
\value{
ExpressionSet
}
\description{
An ExpressionSet objects containing 6 immune subsets (B-cells, CD4,
CD8, Monocytes, Neutrophils, NK) from healthy individuals.
}
\keyword{datasets}
