% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_Panglao_types.R
\name{list_Panglao_types}
\alias{list_Panglao_types}
\title{Show the summary info of available cell types in PanglaoDB.}
\usage{
list_Panglao_types(type = "organ", organ)
}
\arguments{
\item{type}{chr, specify if to list all available organ types or cell types
for specific organ}

\item{organ}{chr, specify the tissue or organ label to list cell types,
optional when type = 'cell'}
}
\value{
a vector of available organ types or cell types in PanglaoDB
}
\description{
Show the name and number of each cell type in PanglaoDB. Help users
know which subset(s) marker list(s) could be retrived by PanglaoDB.
}
\examples{
list_Panglao_types(organ = "Immune system")
}
