% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LM22}
\alias{LM22}
\title{LM22 matrix for CIBERSORT.}
\format{
A data frame with 547 rows 23 variables:
\describe{
\item{Gene}{gene symbols}
\item{B cells naive}{0 or 1, represents if the gene is significantly
up-regulated in the subset}
...
}
}
\source{
\url{https://cibersort.stanford.edu/}
}
\usage{
LM22
}
\description{
A dataset containing 547 marker genes expression of 22 immune subsets which
is generated for CIBERSORT.
}
\keyword{datasets}
