% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LM7}
\alias{LM7}
\title{LM7 matrix for CIBERSORT.}
\format{
A data frame with 375 rows 9 variables:
\describe{
\item{Gene}{gene symbols}
\item{Subset}{immune subset of the marker gene}
\item{B cells}{gene median expression in B cells}
...
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5384348/}
}
\usage{
LM7
}
\description{
A dataset containing 375 marker genes expression of 7 immune subsets which
is generated for CIBERSORT.
}
\keyword{datasets}
