% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msigdb_gobp_nk}
\alias{msigdb_gobp_nk}
\title{Sub-collection of MSigDB gene sets.}
\format{
A GeneSetCollection of 55 gene sets.
}
\source{
\code{\link[msigdb:getMsigdb]{msigdb::getMsigdb()}}
}
\usage{
data(msigdb_gobp_nk)
}
\description{
A small GeneSetCollection object, contains gene sets with gene set name matched
to 'NATURAL_KILLER' from GO:BP MSigDB v7.4 database.
}
\keyword{datasets}
