% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_heatmap.R
\name{sig_heatmap}
\alias{sig_heatmap}
\alias{sig_heatmap,matrix,vector,vector,vector-method}
\alias{sig_heatmap,Matrix,vector,vector,vector-method}
\alias{sig_heatmap,data.frame,vector,vector,vector-method}
\alias{sig_heatmap,DGEList,vector,character,vector-method}
\alias{sig_heatmap,ExpressionSet,vector,character,vector-method}
\alias{sig_heatmap,Seurat,vector,character,vector-method}
\alias{sig_heatmap,SummarizedExperiment,vector,character,vector-method}
\alias{sig_heatmap,list,vector,character,vector-method}
\title{Heatmap original markers and screened signature}
\usage{
sig_heatmap(
  data,
  sigs,
  ID,
  markers,
  counts = TRUE,
  scale = "none",
  min_max = FALSE,
  gene_id = "SYMBOL",
  ranks_plot = FALSE,
  slot = "counts",
  col = colorRampPalette(c("#76B7B2", "#E15759"))(256)
)

\S4method{sig_heatmap}{matrix,vector,vector,vector}(
  data,
  sigs,
  ID,
  markers,
  counts = TRUE,
  scale = "none",
  min_max = FALSE,
  gene_id = "SYMBOL",
  ranks_plot = FALSE,
  col = colorRampPalette(c("#76B7B2", "#E15759"))(256)
)

\S4method{sig_heatmap}{Matrix,vector,vector,vector}(
  data,
  sigs,
  ID,
  markers,
  counts = TRUE,
  scale = "none",
  min_max = FALSE,
  gene_id = "SYMBOL",
  ranks_plot = FALSE,
  col = colorRampPalette(c("#76B7B2", "#E15759"))(256)
)

\S4method{sig_heatmap}{data.frame,vector,vector,vector}(
  data,
  sigs,
  ID,
  markers,
  counts = TRUE,
  scale = "none",
  min_max = FALSE,
  gene_id = "SYMBOL",
  ranks_plot = FALSE,
  col = colorRampPalette(c("#76B7B2", "#E15759"))(256)
)

\S4method{sig_heatmap}{DGEList,vector,character,vector}(
  data,
  sigs,
  ID,
  markers,
  counts = TRUE,
  scale = "none",
  min_max = FALSE,
  gene_id = "SYMBOL",
  ranks_plot = FALSE,
  col = colorRampPalette(c("#76B7B2", "#E15759"))(256)
)

\S4method{sig_heatmap}{ExpressionSet,vector,character,vector}(
  data,
  sigs,
  ID,
  markers,
  counts = TRUE,
  scale = "none",
  min_max = FALSE,
  gene_id = "SYMBOL",
  ranks_plot = FALSE,
  col = colorRampPalette(c("#76B7B2", "#E15759"))(256)
)

\S4method{sig_heatmap}{Seurat,vector,character,vector}(
  data,
  sigs,
  ID,
  markers,
  counts = TRUE,
  scale = "none",
  min_max = FALSE,
  gene_id = "SYMBOL",
  ranks_plot = FALSE,
  slot = "counts",
  col = colorRampPalette(c("#76B7B2", "#E15759"))(256)
)

\S4method{sig_heatmap}{SummarizedExperiment,vector,character,vector}(
  data,
  sigs,
  ID,
  markers,
  counts = TRUE,
  scale = "none",
  min_max = FALSE,
  gene_id = "SYMBOL",
  ranks_plot = FALSE,
  slot = "counts",
  col = colorRampPalette(c("#76B7B2", "#E15759"))(256)
)

\S4method{sig_heatmap}{list,vector,character,vector}(
  data,
  sigs,
  ID,
  markers,
  counts = TRUE,
  scale = "none",
  min_max = FALSE,
  gene_id = "SYMBOL",
  ranks_plot = FALSE,
  slot = "counts",
  col = colorRampPalette(c("#76B7B2", "#E15759"))(256)
)
}
\arguments{
\item{data}{expression data, can be matrix, DGEList, eSet, seurat, sce...}

\item{sigs}{a vector of signature (Symbols)}

\item{ID}{character or vector, specify the column name to compare in coldata}

\item{markers}{a vector of gene names, listed the gene symbols of original
markers pool}

\item{counts}{logical, indicate if data is raw counts data}

\item{scale}{could be 'none', 'row' or 'column'}

\item{min_max}{logical, if to use min_max to normalize data by rows}

\item{gene_id}{character, indicate the ID type of rowname of expression data's ,
could be one of 'ENSEMBL', 'SYMBOL', ... default 'SYMBOL'}

\item{ranks_plot}{logical, if to use ranks instead of expression of genes to
draw heatmap}

\item{slot}{character, indicate which slot used for seurat/sce object}

\item{col}{vactor of color to draw heatmap}
}
\value{
patchwork object of pheatmaps
}
\description{
Compare the heatmap before and after screening.
}
\examples{
data("im_data_6", "NK_markers")
sig_heatmap(data = im_data_6, sigs = NK_markers$HGNC_Symbol[1:10],
            ID = "celltype:ch1", markers = NK_markers$HGNC_Symbol,
            gene_id = "ENSEMBL")

}
