% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_rankdensity_plot.R
\name{sig_rankdensity_plot}
\alias{sig_rankdensity_plot}
\alias{sig_rankdensity_plot,matrix,vector,vector-method}
\alias{sig_rankdensity_plot,Matrix,vector,vector-method}
\alias{sig_rankdensity_plot,data.frame,vector,vector-method}
\alias{sig_rankdensity_plot,DGEList,vector,character-method}
\alias{sig_rankdensity_plot,ExpressionSet,vector,character-method}
\alias{sig_rankdensity_plot,Seurat,vector,character-method}
\alias{sig_rankdensity_plot,SummarizedExperiment,vector,character-method}
\alias{sig_rankdensity_plot,list,vector,character-method}
\title{Plot rank density}
\usage{
sig_rankdensity_plot(
  data,
  sigs,
  ID,
  counts = TRUE,
  aggregate = FALSE,
  slot = "counts",
  gene_id = "SYMBOL"
)

\S4method{sig_rankdensity_plot}{matrix,vector,vector}(
  data,
  sigs,
  ID,
  counts = TRUE,
  aggregate = FALSE,
  gene_id = "SYMBOL"
)

\S4method{sig_rankdensity_plot}{Matrix,vector,vector}(
  data,
  sigs,
  ID,
  counts = TRUE,
  aggregate = FALSE,
  gene_id = "SYMBOL"
)

\S4method{sig_rankdensity_plot}{data.frame,vector,vector}(
  data,
  sigs,
  ID,
  counts = TRUE,
  aggregate = FALSE,
  gene_id = "SYMBOL"
)

\S4method{sig_rankdensity_plot}{DGEList,vector,character}(
  data,
  sigs,
  ID,
  counts = TRUE,
  aggregate = FALSE,
  gene_id = "SYMBOL"
)

\S4method{sig_rankdensity_plot}{ExpressionSet,vector,character}(
  data,
  sigs,
  ID,
  counts = TRUE,
  aggregate = FALSE,
  gene_id = "SYMBOL"
)

\S4method{sig_rankdensity_plot}{Seurat,vector,character}(
  data,
  sigs,
  ID,
  counts = TRUE,
  aggregate = FALSE,
  slot = "counts",
  gene_id = "SYMBOL"
)

\S4method{sig_rankdensity_plot}{SummarizedExperiment,vector,character}(
  data,
  sigs,
  ID,
  counts = TRUE,
  aggregate = FALSE,
  slot = "counts",
  gene_id = "SYMBOL"
)

\S4method{sig_rankdensity_plot}{list,vector,character}(
  data,
  sigs,
  ID,
  counts = TRUE,
  aggregate = FALSE,
  slot = "counts",
  gene_id = "SYMBOL"
)
}
\arguments{
\item{data}{list of immune cells expression data,
default list(im_data_6, im_data_29)}

\item{sigs}{vector of gene symbols to calculate rank scores}

\item{ID}{chr or vector, specify the column name to compare in coldata}

\item{counts}{logical, indicate if data is raw counts data}

\item{aggregate}{logical, if to aggregate expression according to \code{ID},
default FALSE}

\item{slot}{chr, which slot in scdata to use if data is seurat or sce}

\item{gene_id}{chr, indicate the ID type of expression data's rowname,
could be one of 'ENSEMBL', 'SYMBOL', ... default 'SYMBOL'}
}
\value{
ggplot or patchwork
}
\description{
Plot rank density
}
\examples{
sig_rankdensity_plot(
  data = im_data_6, sigs = NK_markers$HGNC_Symbol[1:10],
  ID = "celltype:ch1", gene_id = "ENSEMBL"
)

}
