% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_scatter_plot.R
\name{sig_scatter_plot}
\alias{sig_scatter_plot}
\alias{sig_scatter_plot,matrix,vector,vector,character-method}
\alias{sig_scatter_plot,Matrix,vector,vector,character-method}
\alias{sig_scatter_plot,DGEList,vector,character,character-method}
\alias{sig_scatter_plot,ExpressionSet,vector,character,character-method}
\alias{sig_scatter_plot,Seurat,vector,character,character-method}
\alias{sig_scatter_plot,SummarizedExperiment,vector,character,character-method}
\alias{sig_scatter_plot,list,vector,character,character-method}
\title{Scatter plot of signature for specific subset vs others}
\usage{
sig_scatter_plot(
  data,
  sigs,
  ID,
  type,
  counts = TRUE,
  slot = "counts",
  xint = 1,
  yint = 1,
  gene_id = "SYMBOL"
)

\S4method{sig_scatter_plot}{matrix,vector,vector,character}(
  data,
  sigs,
  ID,
  type,
  counts = TRUE,
  xint = 1,
  yint = 1,
  gene_id = "SYMBOL"
)

\S4method{sig_scatter_plot}{Matrix,vector,vector,character}(
  data,
  sigs,
  ID,
  type,
  counts = TRUE,
  xint = 1,
  yint = 1,
  gene_id = "SYMBOL"
)

\S4method{sig_scatter_plot}{DGEList,vector,character,character}(
  data,
  sigs,
  ID,
  type,
  counts = TRUE,
  xint = 1,
  yint = 1,
  gene_id = "SYMBOL"
)

\S4method{sig_scatter_plot}{ExpressionSet,vector,character,character}(
  data,
  sigs,
  ID,
  type,
  counts = TRUE,
  xint = 1,
  yint = 1,
  gene_id = "SYMBOL"
)

\S4method{sig_scatter_plot}{Seurat,vector,character,character}(
  data,
  sigs,
  ID,
  type,
  counts = TRUE,
  slot = "counts",
  xint = 1,
  yint = 1,
  gene_id = "SYMBOL"
)

\S4method{sig_scatter_plot}{SummarizedExperiment,vector,character,character}(
  data,
  sigs,
  ID,
  type,
  counts = TRUE,
  slot = "counts",
  xint = 1,
  yint = 1,
  gene_id = "SYMBOL"
)

\S4method{sig_scatter_plot}{list,vector,character,character}(
  data,
  sigs,
  ID,
  type,
  counts = TRUE,
  slot = "counts",
  xint = 1,
  yint = 1,
  gene_id = "SYMBOL"
)
}
\arguments{
\item{data}{expression data, can be matrix, DGEList, eSet, seurat, sce...}

\item{sigs}{a vector of signature (Symbols)}

\item{ID}{character or vector, specify the column name to compare in coldata}

\item{type}{pattern, specify the group of interest as reference}

\item{counts}{logical, indicate if data is raw counts data}

\item{slot}{character, indicate which slot used for seurat/sce object}

\item{xint}{intercept of vertical dashed line}

\item{yint}{intercept of horizontal dashed line}

\item{gene_id}{character, indicate the ID type of rowname of expression data's ,
could be one of 'ENSEMBL', 'SYMBOL', ... default 'SYMBOL'}
}
\value{
patchwork or ggplot of scatter plot of median expression
}
\description{
Scatter plot depicts mean expression for each signature gene in
the specific subset against other cell types.
}
\examples{
data("im_data_6", "NK_markers")
sig_scatter_plot(
  sigs = NK_markers$HGNC_Symbol, data = im_data_6,
  ID = "celltype:ch1", type = "NK",
  gene_id = "ENSEMBL"
)

}
